Separate tap-to-focus keys between camera/video
For the camera, the user expects a default tap-to-focus duration of 3 seconds
with exposure locking. For the camcorder, however, the user expects a tap-to-focus
duration of 0 seconds with no exposure locking, as the rapid change in exposure
while recording would negatively impact the video.

Separate the tap-to-focus option to reflect the recommended defaults.

Signed-off-by: Sultanxda <sultanxda@gmail.com>